Add zombie concept to TangoDevice
This is necessary for Sardana test suite which restart the device server
in between the tests. If the GC does not destroy the TangoDevice on time
for the next test then we receive API_EventTimeout errors which then
enable polling and this situation is hard to control in the tests.
With the zombie flag we mark a TangoDevice as disposable and ask the
factory for a new one.